Separate layouts for modules/component definitions
